fix: fusion pass fixes with temporal blocking
